Fixed memory management of stack entries.
Fixed leak of entries when pushing on an empty stack that contained elements
before.
As the elements are freed only when the full stack is freed, we need to keep
a pointer to them. But the way it was done breaks when stack got empty then
filled again, because the first filling cycle got overridden by the new one.
Now it is done through a simple list, fixing the issue.