Switch values to double and long instead of float and ints.
Unify the value types to double and long instead of mixed float, double,
int and long; resulting in a cleaner code.
The cost of this change is a little bigger memory consumption (as long are
generally bigger than int and double are bigger than float), but we gain a
better precision (which is probably not useful tough) and less casts.
Another gain is drop of search for C99's math function working on floats
(like floorf(), roundf(), strtof() and so).

Dropped strtod() in mathutils' ctpl_math_string_to_float() too in favour of
g_ascii_strtod(), for the same reason than in the last commit (locale
insensitivity).

Update the configure accordingly not to check for no longer used functions.