ctpl_lexer_lex() now never return NULL if there was no error.
Rather than returning NULL if the input was completely empty, return
a valid tree with only an empty data token (e.g. the correct representation
of the input).
It is slightly better from the user point of view since he/she doesn't need
to check whether the error was set to know if an error happened, checking
the returned value against NULL is enough.