Improve the multiplication operator.
The multiplication operator no longer behaves differently regarding which
operand is at the left or the right.
The typical case was a string multiplication: it only worked if the
multiplictation amount was on the right; the same kind of thing happened
with mixed integer/reals. The thing was that the operation was always done
on the type of the left operand, and not on the "most appropriated type".