Replace variable statement by expression statement.
Statements with only a variable reference are gone to let place to
new expression statements.
Expression statements are completely backward-compatible with variable
statements since an expression may contain any variable, but they allows
any arbitrary expression where only variable were allowed with variable
statements.
This allows for example '{foo * 2}' statement, that will be replaced by
the value of the 'foo' variable multiplied by 2.
Even if this functionality may not be heavily used in real-world
templates, it was easy to implement and is cleaner regarding the internal
design (no duplication of functionality).