Correctly convert floating-point values to integer ones.
We was checking if the floor()ed value was equal to the original one
to know whether the value is integer in practise, but this way is not
good because C's double can handle widely larger values than C's long int;
then if the floating value had no fraction part and was > to LONG_MAX,
we loosed a part of the value in the cast.
Now, we check whether the value casted to long int is the equal to the
original double value, and then only consider it to be long int compatible.