Make CtplLexerExpr read from a LibMB buffer rather than a string.
Now the expression lexer reads directly from a MB.
This change brakes the API: remember to update your code. An easy upgrade
(but probably not the better one) is to use the new
ctpl_lexer_expr_lex_string() that behaves the same way as the old
ctpl_lexer_expr_lex().