Speedup large tree manipulation
Add pseudo-end and start pointers to each token so they (hopefully)
don't have to traverse the list to append or prepend neighbours.
This implementation is an easy fix but doesn't really speeds up
appending or prepending from the middle of the tree, but only from
the bounds.
However, it is actually an efficient fix since we never append or
prepend from the middle of a tree.