Speed improvements for CtplStack
Always add elements at start of lists rather than at the end not to
have to walk down the list to get the last element to append to.
Also free the list and its elements at once rather than walking the
list twice.

This speed up a lot parsing when there is a symbol that changes often.