Encoding support in CLI tool is no longer optional
Having it optional makes the code harder to read/maintain, and makes
possible to have two different version of the same program behaving
a little diferrently, which is a source of confusion and a pain to
maintain.

This makes the CLI tool depend on GLib 2.24 (to have
GConverter{In,Out}putStream as well as GCharsetConverter).