Fix output of floating point numbers
Floating point numbers were merely unusable because of the C's double
precision that is not perfect. The problem was that for example "3.14"
was output with garbage at the end (actually "3.1400000000000001")
because of that imprecision -- which ironically lead to unwanted
precision.
Even if the value is the real value stored into memory, this is
meaningless in templates and makes floats not good-looking at all, and
so merely unusable.

Technically, the fix is simply to reduce a little bit the output
precision, which seems to give good results -- and without too much
precision loss, 15 digits vs. 17. This change may modify the output of
a template (as seen as a character sequence) but not its meaning,
since the removed "precision" was most likely to be unwanted.