Don't eat characters that will not be part of the final number
Don't eat characters that are not actually part of the finial number.
For example, don't eat the x of "0xzzz" since only the 0 by itself is
a valid number, z not being a valid hexadecimal digit.
This is not useful yet since any of these extra characters will cause
an error later on the lexing, but it would allow to have name clashes
between operators and numbers without problem.

Also refactored the code a bit to be more consistent.

We also unexpectedly won about 2% of performances, cool.