Implement array comparison
Implement booleans ==, !=, <=, <, > and >= for arrays. This comparison
compares each element of the array to get the final comparison.

This commit also changes a bit the implementation of the comparison operators,
now internally using a strcmp-like value to represent the difference (< 0 if
inferior, = 0 is equal, > 0 if superior).

Finally, this commit fixes a crash when trying to compare an array with
something else. This wasn't implemented but returned a value indicating an
error, but that error wasn't set, leading to a crash. Fixed together with
the implementation.