Testsuite: don't rely on the CLI tool to check the library's behavior
The CLI tool depends on a newer version of GIO and on GIO-UNIX than the
library, so it is not correct to depend on it for the library's tests.

So, add a C test program that does what tests.sh did, but using the
library directly.

For now tests.sh is kept untouched, but only run if the CLI tool is
built. In the future it should test the CLI tool behavior rather than
the library's behavior.