Improve filter operator priority
Make the filter operator have the same priority as the other lowest
priority operators.  This allows to behave more naturally when in the
middle of an expression, like this:

	foo + bar | filter + baz

The previous priority, which was lower than the addition operator,
resulted in:

	(foo + bar) | (filter + baz)

While the left side is good, the right one isn't as the expression is
invalid (operator + can't work on filters).

New behavior gives addition and filter operators the same priority, and
so ordering is taken into account, which leads to interpreting as:

	((foo + bar) | filter) + baz

Which makes more sense as it is valid.