Add explicit cast for recent GLib and GCC
Recent versions of GLib with recent enough versions of GCC will have
g_object_ref() return type be the same as the argument type instead of
a generic pointer.

This thus lead to a incompatible pointer type in the return statement,
which GCC 14 doesn't seem to like so much.

Just add an explicit cast here as there's nothing more to it.

Fixes https://github.com/b4n/ctpl/issues/2